@@ -16,73 +16,78 @@
     "prepare": "husky"
   },
   "dependencies": {
-    "@hookform/resolvers": "^2.9.11",
-    "@neondatabase/serverless": "^0.9.0",
-    "@paypal/paypal-js": "^5.1.5",
-    "@paypal/react-paypal-js": "^7.8.2",
-    "@prisma/adapter-neon": "^5.12.1",
-    "@prisma/client": "5.12.1",
-    "@prisma/extension-accelerate": "^1.0.0",
+    "@hookform/resolvers": "^3.9.1",
+    "@neondatabase/serverless": "^0.9.5",
+    "@paypal/paypal-js": "^5.1.6",
+    "@paypal/react-paypal-js": "^7.8.3",
+    "@prisma/adapter-neon": "^6.1.0",
+    "@prisma/client": "6.1.0",
+    "@prisma/extension-accelerate": "^1.2.1",
+    "@radix-ui/react-slot": "^1.1.1",
     "@sendgrid/mail": "^7.7.0",
-    "@tanstack/match-sorter-utils": "^8.8.4",
-    "@tanstack/react-query": "^4.28.0",
-    "@tanstack/react-table": "^8.8.4",
+    "@tanstack/match-sorter-utils": "^8.19.4",
+    "@tanstack/react-form": "^0.40.4",
+    "@tanstack/react-query": "^4.36.1",
+    "@tanstack/react-table": "^8.20.6",
     "@trpc/client": "^10.45.2",
     "@trpc/next": "^10.45.2",
     "@trpc/react-query": "^10.45.2",
     "@trpc/server": "^10.45.2",
     "@vercel/analytics": "^0.1.11",
-    "@vercel/speed-insights": "^1.0.1",
-    "csv": "^6.3.5",
+    "@vercel/speed-insights": "^1.1.0",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "csv": "^6.3.11",
     "immer": "^9.0.21",
-    "jotai": "^2.0.3",
+    "jotai": "^2.10.4",
     "jotai-devtools": "^0.2.0",
-    "next": "^13.5.5",
-    "next-auth": "^4.20.1",
-    "react": "18.2.0",
-    "react-datepicker": "^4.10.0",
-    "react-dom": "18.2.0",
-    "react-hook-form": "^7.43.8",
+    "next": "^15.1.2",
+    "next-auth": "^4.24.11",
+    "react": "19.0.0",
+    "react-datepicker": "^7.5.0",
+    "react-dom": "19.0.0",
+    "react-hook-form": "^7.54.2",
     "react-query": "^3.39.3",
-    "superjson": "^1.12.2",
-    "ws": "^8.16.0",
-    "zod": "3.23.8",
-    "zod-validation-error": "^3.3.0"
+    "superjson": "^1.13.3",
+    "tailwind-merge": "^2.5.5",
+    "type-fest": "^4.30.2",
+    "ws": "^8.18.0",
+    "zod": "3.24.1",
+    "zod-validation-error": "^3.4.0"
   },
   "devDependencies": {
     "@faker-js/faker": "^7.6.0",
-    "@next/bundle-analyzer": "^13.2.4",
+    "@next/bundle-analyzer": "^13.5.8",
     "@swc-jotai/react-refresh": "^0.0.7",
-    "@types/jest": "^29.5.0",
-    "@types/node": "^18.15.10",
-    "@types/react": "^18.0.29",
-    "@types/react-datepicker": "^4.10.0",
-    "@types/react-dom": "18.0.11",
-    "@types/ws": "^8.5.10",
-    "@typescript-eslint/eslint-plugin": "^6.10.0",
-    "@typescript-eslint/parser": "^6.10.0",
-    "autoprefixer": "^10.4.16",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.10.2",
+    "@types/react": "^19.0.2",
+    "@types/react-dom": "19.0.2",
+    "@types/ws": "^8.5.13",
+    "@typescript-eslint/eslint-plugin": "^8.18.1",
+    "@typescript-eslint/parser": "^8.18.1",
+    "autoprefixer": "^10.4.20",
     "bufferutil": "^4.0.8",
-    "daisyui": "^2.51.5",
-    "dotenv-cli": "^7.1.0",
-    "eslint": "8.35.0",
-    "eslint-config-next": "13.2.3",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.33.2",
-    "husky": "^9.0.11",
-    "jest": "^29.5.0",
-    "jest-mock-extended": "^3.0.4",
-    "lint-staged": "^15.2.2",
-    "postcss": "^8.4.31",
-    "prettier": "^2.8.7",
-    "prettier-plugin-tailwindcss": "^0.2.5",
-    "prisma": "5.12.1",
-    "tailwindcss": "^3.3.5",
-    "ts-jest": "^29.0.5",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.4.0",
+    "daisyui": "^2.52.0",
+    "dotenv-cli": "^7.4.4",
+    "eslint": "9.17.0",
+    "eslint-config-next": "15.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.37.2",
+    "husky": "^9.1.7",
+    "jest": "^29.7.0",
+    "jest-mock-extended": "^3.0.7",
+    "lint-staged": "^15.2.11",
+    "postcss": "^8.4.49",
+    "prettier": "^3.4.2",
+    "prettier-plugin-tailwindcss": "^0.6.9",
+    "prisma": "6.1.0",
+    "tailwindcss": "^3.4.17",
+    "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.2",
     "utf-8-validate": "^6.0.5",
-    "zod-prisma-types": "3.1.8"
+    "zod-prisma-types": "3.2.1"
   },
   "prisma": {
     "schema": "./src/server/prisma/schema.prisma"
