@@ -1,5 +1,5 @@
 import { PropsWithChildren } from 'react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import { useAtom } from 'jotai';
 import {
   CreateOrderActions,
@@ -11,6 +11,11 @@ import {
 import { formState } from '@/utils/atoms';
 import Alert from './Alert';
 import PayPalButton from '../styled-ui/PayPalButton';
+import { Button } from '../styled-ui/Button';
+import {
+  PayPalScriptProvider,
+  ReactPayPalScriptOptions,
+} from '@paypal/react-paypal-js';
 
 interface PaymentProps extends PropsWithChildren {
   showPayment: boolean;
@@ -20,6 +25,14 @@ interface PaymentProps extends PropsWithChildren {
     mutateFn: () => void;
   };
 }
+
+const initOptions: ReactPayPalScriptOptions = {
+  'client-id':
+    process.env.NEXT_PUBLIC_PAYPAL_CLIENT_ID || process.env.PAYPAL_CLIENT_ID,
+  currency: 'USD',
+  intent: 'capture',
+  'data-react-paypal-script-id': 'paypal-button',
+};
 
 function Payment({ showPayment, children, formMutation }: PaymentProps) {
   const history = useRouter();
@@ -56,31 +69,33 @@ function Payment({ showPayment, children, formMutation }: PaymentProps) {
 
   return (
     <>
-      {showPayment ? (
-        <div className='grid place-content-center gap-5 rounded-2xl border border-solid border-gray-400 bg-gray-100 p-5'>
-          <h4 className='text-lg font-bold'>
-            Your sign-up costs ${amountOwed} USD
-          </h4>
-          <PayPalButton
-            createOrder={createOrder}
-            onApprove={onApprove}
-          />
-        </div>
-      ) : (
-        <>
-          {children}
-          <Alert
-            message={formMutation.message ?? ''}
-            visible={formMutation.error}
-          />
-          <button
-            type='submit'
-            className='btn btn-primary w-full'
-          >
-            Move to payment
-          </button>
-        </>
-      )}
+      <PayPalScriptProvider options={initOptions}>
+        {showPayment ? (
+          <div className='grid place-content-center gap-5 rounded-2xl border border-solid border-gray-400 bg-gray-100 p-5'>
+            <h4 className='text-lg font-bold'>
+              Your sign-up costs ${amountOwed} USD
+            </h4>
+            <PayPalButton
+              createOrder={createOrder}
+              onApprove={onApprove}
+            />
+          </div>
+        ) : (
+          <>
+            {children}
+            <Alert
+              message={formMutation.message ?? ''}
+              visible={formMutation.error}
+            />
+            <Button
+              type='submit'
+              className='w-full'
+            >
+              Move to payment
+            </Button>
+          </>
+        )}
+      </PayPalScriptProvider>
     </>
   );
 }
