@@ -1,14 +1,20 @@
 'use client';
 import { usePathname, useRouter } from 'next/navigation';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import Radio from '@/components/styled-ui/Radio';
 import { FORMTYPE } from '@/types/common';
+import { Button } from '@/components/styled-ui/Button';
 
 function FormSelector() {
   const router = useRouter();
   const path = usePathname();
   const [type, setType] = useState<string>();
+  useEffect(() => {
+    if (type) {
+      router.prefetch(`${path}/${type}`);
+    }
+  }, [router, path, type]);
 
   return (
     <>
@@ -21,14 +27,14 @@ function FormSelector() {
             <label
               htmlFor={`${selection}-id`}
               aria-label={selection}
-              className='label cursor-pointer space-x-5'
+              className='label cursor-pointer space-x-5 capitalize'
               key={selection}
             >
               <Radio
                 id={`${selection}-id`}
-                className='radio radio-primary radio-md md:radio-sm'
+                name='form-select'
+                className='radio-md md:radio-sm'
                 value={selection}
-                name='app-select'
                 onClick={() => setType(selection)}
               />
               <span className='label-text'>{selection}</span>
@@ -36,16 +42,16 @@ function FormSelector() {
           </div>
         ))}
       </div>
-      <button
-        className='btn btn-primary w-full'
+      <Button
+        className='w-full'
         onClick={() => {
           if (type) {
-            router.push(`${path}/${type.toLowerCase()}`);
+            router.push(`${path}/${type}`);
           }
         }}
       >
         Next
-      </button>
+      </Button>
     </>
   );
 }
