@@ -1,62 +1,69 @@
+'use client';
 import { useSetAtom } from 'jotai';
-import { ReactElement, useState } from 'react';
+import { useState, useTransition } from 'react';
 
-import { FormLayout } from '@/components/layout/FormLayout';
 import { MemberForm, MemberFormSchema } from '@/utils/zodschemas';
 import { updateFormState } from '@/utils/atoms';
 import { capitalize } from '@/utils/helpers';
 import states from '@/utils/states.json';
-import { trpc } from '@/utils/trpc';
 import useZodForm from '@/utils/usezodform';
 import Form from '@/components/forms/Form';
 import { PhoneTypeSchema } from '@/server/prisma/zod-generated/inputTypeSchemas/PhoneTypeSchema';
 import Payment from '@/components/forms/Payment';
 import Input from '@/components/data-entry/Input';
 import Select from '@/components/data-entry/Select';
-import RegType from '@/components/forms/RegType';
+import RegistrationSelect from '@/components/forms/RegType';
 import MemberType from '@/components/forms/MemberType';
 import Under18 from '@/components/forms/Under18';
 import Checkbox from '@/components/data-entry/Checkbox';
 import HorseFieldArray from '@/components/forms/HorseFieldArray';
 import RegistrationYearSelect from '@/components/forms/RegistrationYearSelect';
+import { checkForExistingMember } from '@/app/action';
+import { Label } from '@/components/styled-ui/Label';
+
+export interface ActionState {
+  message: string | undefined;
+  error: boolean;
+  data: MemberForm | undefined;
+}
+
+export const initialState: ActionState = {
+  message: undefined,
+  error: false,
+  data: undefined,
+};
 
 function IndividualRegistration() {
   const [payment, togglePayment] = useState(false);
   const [isRegHorse, toggleRegHorse] = useState(false);
   const update = useSetAtom(updateFormState);
 
-  const checkMember = trpc.members.exists.useMutation({
-    onSuccess() {
-      togglePayment(true);
-    },
-  });
+  const [pending, startTransition] = useTransition();
 
-  const add = trpc.members.add.useMutation();
+  const add: any = () => {};
 
   const form = useZodForm({
     reValidateMode: 'onSubmit',
     shouldFocusError: true,
     schema: MemberFormSchema,
     defaultValues: {
-      memberInput: {
-        businessName: null,
-        dateOfBirth: null,
-        memberType: 'Individual',
-        membershipEnd: null,
-      },
+      dateOfBirth: null,
+      memberType: 'Individual',
+      membershipEnd: null,
     },
   });
 
   const { register, control } = form;
+  const [actionState, setActionState] = useState(initialState);
 
   function onSubmit(formValues: MemberForm) {
     if (formValues.horses) {
       const lifeCount = formValues.horses.filter(
-        horse => horse.regType === 'Life'
+        horse => horse.regType === 'Life',
       ).length;
 
       const annualCount = formValues.horses.filter(
-        horse => horse.regType === 'Annual'
+        horse => horse.regType === 'Annual',
       ).length;
 
       update({
@@ -65,8 +72,16 @@ function IndividualRegistration() {
       });
     }
 
-    checkMember.mutate(formValues);
+    startTransition(async () => {
+      const test = await checkForExistingMember(formValues);
+      setActionState(test);
+      if (!test.error) {
+        togglePayment(curr => !curr);
+      }
+    });
   }
+
+  console.log(pending);
 
   function handleRegHorseCheck() {
     toggleRegHorse(curr => !curr);
@@ -80,8 +95,8 @@ function IndividualRegistration() {
       <Payment
         showPayment={payment}
         formMutation={{
-          error: checkMember.isError,
-          message: checkMember.error?.message,
+          error: actionState.error,
+          message: actionState.message,
           mutateFn: () => add.mutate(form.getValues()),
         }}
       >
@@ -91,30 +106,29 @@ function IndividualRegistration() {
           <Input
             type='text'
             label='First Name*'
-            className='input input-bordered input-primary w-full md:input-sm'
-            {...register('memberInput.firstName')}
+            {...register('firstName')}
           />
 
           <Input
             type='text'
             label='Last Name*'
-            className='input input-bordered input-primary w-full md:input-sm'
-            {...register('memberInput.lastName')}
+            {...register('lastName')}
           />
         </div>
 
-        <h3 className='mt-3 pb-2 text-sm'>Address*</h3>
-        <div className='flex flex-col gap-2'>
+        <Label htmlFor='address-group'>Address*</Label>
+        <div
+          id='address-group'
+          className='flex flex-col gap-2'
+        >
           <Input
             type='text'
-            className='input input-bordered input-primary w-full md:input-sm'
             placeholder='Address Line 1'
-            {...register('memberInput.address')}
+            {...register('address')}
           />
 
           <Input
             type='text'
-            className='input input-bordered input-primary w-full md:input-sm'
             placeholder='Address Line 2'
             name='temp'
           />
@@ -122,15 +136,11 @@ function IndividualRegistration() {
           <div className='flex flex-col gap-1 md:flex-row'>
             <Input
               type='text'
-              className='input input-bordered input-primary w-full md:input-sm'
               placeholder='City'
-              {...register('memberInput.city')}
+              {...register('city')}
             />
 
-            <Select
-              className='select select-bordered select-primary w-full md:select-sm'
-              {...register('memberInput.state')}
-            >
+            <Select {...register('state')}>
               {states.map(state => (
                 <option
                   key={state.value}
@@ -143,9 +153,8 @@ function IndividualRegistration() {
 
             <Input
               type='numeric'
-              className='input input-bordered input-primary w-full md:input-sm'
               placeholder='Zip Code'
-              {...register('memberInput.zip', { valueAsNumber: true })}
+              {...register('zip', { valueAsNumber: true })}
             />
           </div>
 
@@ -153,8 +162,7 @@ function IndividualRegistration() {
             <div className='flex gap-2'>
               <Select
                 label='Phone Type*'
-                className='select select-bordered select-primary md:select-sm'
-                {...register('memberInput.phoneType')}
+                {...register('phoneType')}
               >
                 {Object.keys(PhoneTypeSchema.enum).map(type => (
                   <option
@@ -169,41 +177,38 @@ function IndividualRegistration() {
               <Input
                 label='Phone Number*'
                 type='tel'
-                className='input input-bordered input-primary w-full md:input-sm'
-                {...register('memberInput.phone')}
+                {...register('phone')}
               />
             </div>
 
             <Input
               label='Email*'
               type='text'
-              className='input input-bordered input-primary w-full md:input-sm'
               altLabel={'This will be the primary method of contact.'}
-              {...register('memberInput.email')}
+              {...register('email')}
             />
           </div>
 
-          <div className='container flex-col'>
-            <RegType
-              register={register('memberInput.memberStatus')}
-              formType='Individual'
+          <div className='flex-col'>
+            <RegistrationSelect
+              register={register('memberStatus')}
+              formType='individual'
             />
 
             <RegistrationYearSelect
               heading='Which year are you registering for?'
-              watchFieldName='memberInput.memberStatus'
+              watchFieldName='memberStatus'
               control={control}
-              register={register('memberInput.membershipEnd')}
+              register={register('membershipEnd')}
             />
 
-            <MemberType register={register('memberInput.memberStatusType')} />
+            <MemberType register={register('memberStatusType')} />
 
-            <Under18 dateName='memberInput.dateOfBirth' />
+            <Under18 dateName='dateOfBirth' />
           </div>
 
           <Checkbox
             label='Do you plan to register your horse(s)?'
-            className='checkbox checkbox-primary md:checkbox-sm'
             checked={isRegHorse}
             onChange={handleRegHorseCheck}
           />
@@ -214,9 +219,5 @@ function IndividualRegistration() {
     </Form>
   );
 }
-
-IndividualRegistration.getLayout = (page: ReactElement) => {
-  return <FormLayout>{page}</FormLayout>;
-};
 
 export default IndividualRegistration;
