@@ -1,4 +1,4 @@
-import { join } from '@/utils/helpers';
+import { cn } from '@/utils/helpers';
 import { forwardRef } from 'react';
 import type { ReactNode, DetailedHTMLProps, InputHTMLAttributes } from 'react';
 
@@ -25,9 +25,9 @@ const Checkbox = forwardRef<HTMLInputElement, CheckboxProps>((props, ref) => (
           ref={ref}
           id='checkbox-input'
           type='checkbox'
-          className={join(
+          className={cn(
             'checkbox checkbox-primary md:checkbox-sm',
-            props.className
+            props.className,
           )}
           {...props}
         />
@@ -37,9 +37,9 @@ const Checkbox = forwardRef<HTMLInputElement, CheckboxProps>((props, ref) => (
         ref={ref}
         id='checkbox-input'
         type='checkbox'
-        className={join(
+        className={cn(
           'checkbox checkbox-primary md:checkbox-sm',
-          props.className
+          props.className,
         )}
         {...props}
       />
