@@ -1,10 +1,11 @@
 import { useFieldArray, useFormContext } from 'react-hook-form';
 import { Prisma, Status } from '@prisma/client';
-import RegType from './RegType';
+import RegistrationSelect from './RegType';
 import { AddIcon, TrashIcon } from '../icons';
 import Input from '../data-entry/Input';
 import RegistrationYearSelect from './RegistrationYearSelect';
 import { setMembershipYear } from '@/server/router/utils';
+import { Button } from '../styled-ui/Button';
 
 type Horses = {
   horses: Array<Prisma.HorseCreateManyInput>;
@@ -48,19 +49,18 @@ export default function HorseFieldArray() {
           </h2>
 
           <div className='card-body'>
-            <RegType
-              noAtomUpdate={true}
+            <RegistrationSelect
               register={register(`horses.${index}.regType` as const, {
                 required: true,
               })}
-              formType='Horse'
+              formType='horse'
               onClick={event => {
                 if (event.currentTarget.value === null) {
                   setValue(`horses.${index}.registrationEnd`, null);
                 } else {
                   setValue(
                     `horses.${index}.registrationEnd`,
-                    setMembershipYear()
+                    setMembershipYear(),
                   );
                 }
               }}
@@ -73,24 +73,20 @@ export default function HorseFieldArray() {
               register={register(`horses.${index}.registrationEnd` as const)}
             />
 
-            <span className='flex flex-col gap-2'>
+            <div className='flex flex-col gap-2'>
               <Input
                 label='Registered Name*'
                 type='text'
-                altLabel='Horses registered name must be used when entering a show.'
-                className='input input-bordered input-primary input-sm w-full'
-                {...register(`horses.${index}.horseRN` as const, {
-                  required: true,
-                })}
+                altLabel='Registered name must be used when entering a show.'
+                {...register(`horses.${index}.horseRN` as const)}
               />
 
               <Input
                 label='Barn Name'
                 type='text'
-                className='input input-bordered input-primary input-sm w-full'
                 {...register(`horses.${index}.horseAKA` as const)}
               />
-            </span>
+            </div>
           </div>
         </div>
       ))}
@@ -98,9 +94,8 @@ export default function HorseFieldArray() {
       <p className='text-xl font-semibold text-error'>
         {errors.horses?.message}
       </p>
-      <button
-        className='btn btn-secondary btn-xs mb-5 w-full'
-        type='button'
+      <Button
+        className='btn-secondary btn-xs mb-5'
         onClick={() =>
           append({
             horseRN: '',
@@ -111,7 +106,7 @@ export default function HorseFieldArray() {
         }
       >
         {AddIcon} Add Horse
-      </button>
+      </Button>
     </section>
   );
 }
