@@ -3,9 +3,9 @@ import type { Prisma } from '@prisma/client';
 import { z } from 'zod';
 import { HorseWhereInputSchema } from './HorseWhereInputSchema';
 
-export const HorseRelationFilterSchema: z.ZodType<Prisma.HorseRelationFilter> = z.object({
+export const HorseScalarRelationFilterSchema: z.ZodType<Prisma.HorseScalarRelationFilter> = z.object({
   is: z.lazy(() => HorseWhereInputSchema).optional(),
   isNot: z.lazy(() => HorseWhereInputSchema).optional()
 }).strict();
 
-export default HorseRelationFilterSchema;
+export default HorseScalarRelationFilterSchema;
