@@ -3,9 +3,9 @@ import type { Prisma } from '@prisma/client';
 import { z } from 'zod';
 import { StringFilterSchema } from './StringFilterSchema';
 import { FloatFilterSchema } from './FloatFilterSchema';
-import { RiderComboRelationFilterSchema } from './RiderComboRelationFilterSchema';
+import { RiderComboScalarRelationFilterSchema } from './RiderComboScalarRelationFilterSchema';
 import { RiderComboWhereInputSchema } from './RiderComboWhereInputSchema';
-import { ShowRelationFilterSchema } from './ShowRelationFilterSchema';
+import { ShowScalarRelationFilterSchema } from './ShowScalarRelationFilterSchema';
 import { ShowWhereInputSchema } from './ShowWhereInputSchema';
 
 export const PointsWhereInputSchema: z.ZodType<Prisma.PointsWhereInput> = z.object({
@@ -17,8 +17,8 @@ export const PointsWhereInputSchema: z.ZodType<Prisma.PointsWhereInput> = z.obje
   points: z.union([ z.lazy(() => FloatFilterSchema),z.number() ]).optional(),
   place: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
   showUid: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
-  RiderCombo: z.union([ z.lazy(() => RiderComboRelationFilterSchema),z.lazy(() => RiderComboWhereInputSchema) ]).optional(),
-  show: z.union([ z.lazy(() => ShowRelationFilterSchema),z.lazy(() => ShowWhereInputSchema) ]).optional(),
+  RiderCombo: z.union([ z.lazy(() => RiderComboScalarRelationFilterSchema),z.lazy(() => RiderComboWhereInputSchema) ]).optional(),
+  show: z.union([ z.lazy(() => ShowScalarRelationFilterSchema),z.lazy(() => ShowWhereInputSchema) ]).optional(),
 }).strict();
 
 export default PointsWhereInputSchema;
