@@ -9,9 +9,9 @@ import { FloatFilterSchema } from './FloatFilterSchema';
 import { IntFilterSchema } from './IntFilterSchema';
 import { BoolFilterSchema } from './BoolFilterSchema';
 import { IntNullableFilterSchema } from './IntNullableFilterSchema';
-import { MemberRelationFilterSchema } from './MemberRelationFilterSchema';
+import { MemberScalarRelationFilterSchema } from './MemberScalarRelationFilterSchema';
 import { MemberWhereInputSchema } from './MemberWhereInputSchema';
-import { HorseRelationFilterSchema } from './HorseRelationFilterSchema';
+import { HorseScalarRelationFilterSchema } from './HorseScalarRelationFilterSchema';
 import { HorseWhereInputSchema } from './HorseWhereInputSchema';
 import { PointsListRelationFilterSchema } from './PointsListRelationFilterSchema';
 import { ShowListRelationFilterSchema } from './ShowListRelationFilterSchema';
@@ -31,8 +31,8 @@ export const RiderComboWhereInputSchema: z.ZodType<Prisma.RiderComboWhereInput> 
   memberName: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
   horseName: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
   showYear: z.union([ z.lazy(() => IntNullableFilterSchema),z.number() ]).optional().nullable(),
-  member: z.union([ z.lazy(() => MemberRelationFilterSchema),z.lazy(() => MemberWhereInputSchema) ]).optional(),
-  horse: z.union([ z.lazy(() => HorseRelationFilterSchema),z.lazy(() => HorseWhereInputSchema) ]).optional(),
+  member: z.union([ z.lazy(() => MemberScalarRelationFilterSchema),z.lazy(() => MemberWhereInputSchema) ]).optional(),
+  horse: z.union([ z.lazy(() => HorseScalarRelationFilterSchema),z.lazy(() => HorseWhereInputSchema) ]).optional(),
   points: z.lazy(() => PointsListRelationFilterSchema).optional(),
   shows: z.lazy(() => ShowListRelationFilterSchema).optional()
 }).strict();
