@@ -11,9 +11,9 @@ import { IntFilterSchema } from './IntFilterSchema';
 import { BoolFilterSchema } from './BoolFilterSchema';
 import { StringFilterSchema } from './StringFilterSchema';
 import { IntNullableFilterSchema } from './IntNullableFilterSchema';
-import { MemberRelationFilterSchema } from './MemberRelationFilterSchema';
+import { MemberScalarRelationFilterSchema } from './MemberScalarRelationFilterSchema';
 import { MemberWhereInputSchema } from './MemberWhereInputSchema';
-import { HorseRelationFilterSchema } from './HorseRelationFilterSchema';
+import { HorseScalarRelationFilterSchema } from './HorseScalarRelationFilterSchema';
 import { HorseWhereInputSchema } from './HorseWhereInputSchema';
 import { PointsListRelationFilterSchema } from './PointsListRelationFilterSchema';
 import { ShowListRelationFilterSchema } from './ShowListRelationFilterSchema';
@@ -46,8 +46,8 @@ export const RiderComboWhereUniqueInputSchema: z.ZodType<Prisma.RiderComboWhereU
   memberName: z.union([ z.lazy(() => StringFilterSchema),z.string().trim().min(1, { message: "Member Name is required" }) ]).optional(),
   horseName: z.union([ z.lazy(() => StringFilterSchema),z.string().trim().min(1, { message: "Horse Name is required" }) ]).optional(),
   showYear: z.union([ z.lazy(() => IntNullableFilterSchema),z.number().int() ]).optional().nullable(),
-  member: z.union([ z.lazy(() => MemberRelationFilterSchema),z.lazy(() => MemberWhereInputSchema) ]).optional(),
-  horse: z.union([ z.lazy(() => HorseRelationFilterSchema),z.lazy(() => HorseWhereInputSchema) ]).optional(),
+  member: z.union([ z.lazy(() => MemberScalarRelationFilterSchema),z.lazy(() => MemberWhereInputSchema) ]).optional(),
+  horse: z.union([ z.lazy(() => HorseScalarRelationFilterSchema),z.lazy(() => HorseWhereInputSchema) ]).optional(),
   points: z.lazy(() => PointsListRelationFilterSchema).optional(),
   shows: z.lazy(() => ShowListRelationFilterSchema).optional()
 }).strict());
