@@ -118,16 +118,16 @@ export { BoardmemberCountOrderByAggregateInputSchema } from './BoardmemberCountO
 export { BoardmemberMaxOrderByAggregateInputSchema } from './BoardmemberMaxOrderByAggregateInputSchema';
 export { BoardmemberMinOrderByAggregateInputSchema } from './BoardmemberMinOrderByAggregateInputSchema';
 export { EnumPositionWithAggregatesFilterSchema } from './EnumPositionWithAggregatesFilterSchema';
-export { NonMemberHorseOwnerNullableRelationFilterSchema } from './NonMemberHorseOwnerNullableRelationFilterSchema';
-export { MemberNullableRelationFilterSchema } from './MemberNullableRelationFilterSchema';
+export { NonMemberHorseOwnerNullableScalarRelationFilterSchema } from './NonMemberHorseOwnerNullableScalarRelationFilterSchema';
+export { MemberNullableScalarRelationFilterSchema } from './MemberNullableScalarRelationFilterSchema';
 export { HorseCountOrderByAggregateInputSchema } from './HorseCountOrderByAggregateInputSchema';
 export { HorseMaxOrderByAggregateInputSchema } from './HorseMaxOrderByAggregateInputSchema';
 export { HorseMinOrderByAggregateInputSchema } from './HorseMinOrderByAggregateInputSchema';
 export { EnumDivisionFilterSchema } from './EnumDivisionFilterSchema';
 export { FloatFilterSchema } from './FloatFilterSchema';
 export { IntNullableFilterSchema } from './IntNullableFilterSchema';
-export { MemberRelationFilterSchema } from './MemberRelationFilterSchema';
-export { HorseRelationFilterSchema } from './HorseRelationFilterSchema';
+export { MemberScalarRelationFilterSchema } from './MemberScalarRelationFilterSchema';
+export { HorseScalarRelationFilterSchema } from './HorseScalarRelationFilterSchema';
 export { PointsListRelationFilterSchema } from './PointsListRelationFilterSchema';
 export { ShowListRelationFilterSchema } from './ShowListRelationFilterSchema';
 export { PointsOrderByRelationAggregateInputSchema } from './PointsOrderByRelationAggregateInputSchema';
@@ -141,8 +141,8 @@ export { RiderComboSumOrderByAggregateInputSchema } from './RiderComboSumOrderBy
 export { EnumDivisionWithAggregatesFilterSchema } from './EnumDivisionWithAggregatesFilterSchema';
 export { FloatWithAggregatesFilterSchema } from './FloatWithAggregatesFilterSchema';
 export { IntNullableWithAggregatesFilterSchema } from './IntNullableWithAggregatesFilterSchema';
-export { RiderComboRelationFilterSchema } from './RiderComboRelationFilterSchema';
-export { ShowRelationFilterSchema } from './ShowRelationFilterSchema';
+export { RiderComboScalarRelationFilterSchema } from './RiderComboScalarRelationFilterSchema';
+export { ShowScalarRelationFilterSchema } from './ShowScalarRelationFilterSchema';
 export { PointsCountOrderByAggregateInputSchema } from './PointsCountOrderByAggregateInputSchema';
 export { PointsAvgOrderByAggregateInputSchema } from './PointsAvgOrderByAggregateInputSchema';
 export { PointsMaxOrderByAggregateInputSchema } from './PointsMaxOrderByAggregateInputSchema';
