@@ -1,23 +1,11 @@
 import { z } from 'zod';
 import { StatusSchema } from '../inputTypeSchemas/StatusSchema'
-import type { NonMemberHorseOwnerWithRelations } from './NonMemberHorseOwnerSchema'
-import type { NonMemberHorseOwnerPartialWithRelations } from './NonMemberHorseOwnerSchema'
-import type { NonMemberHorseOwnerOptionalDefaultsWithRelations } from './NonMemberHorseOwnerSchema'
-import type { MemberWithRelations } from './MemberSchema'
-import type { MemberPartialWithRelations } from './MemberSchema'
-import type { MemberOptionalDefaultsWithRelations } from './MemberSchema'
-import type { RiderComboWithRelations } from './RiderComboSchema'
-import type { RiderComboPartialWithRelations } from './RiderComboSchema'
-import type { RiderComboOptionalDefaultsWithRelations } from './RiderComboSchema'
-import { NonMemberHorseOwnerWithRelationsSchema } from './NonMemberHorseOwnerSchema'
-import { NonMemberHorseOwnerPartialWithRelationsSchema } from './NonMemberHorseOwnerSchema'
-import { NonMemberHorseOwnerOptionalDefaultsWithRelationsSchema } from './NonMemberHorseOwnerSchema'
-import { MemberWithRelationsSchema } from './MemberSchema'
-import { MemberPartialWithRelationsSchema } from './MemberSchema'
-import { MemberOptionalDefaultsWithRelationsSchema } from './MemberSchema'
-import { RiderComboWithRelationsSchema } from './RiderComboSchema'
-import { RiderComboPartialWithRelationsSchema } from './RiderComboSchema'
-import { RiderComboOptionalDefaultsWithRelationsSchema } from './RiderComboSchema'
+import { NonMemberHorseOwnerWithRelationsSchema, NonMemberHorseOwnerPartialWithRelationsSchema, NonMemberHorseOwnerOptionalDefaultsWithRelationsSchema } from './NonMemberHorseOwnerSchema'
+import type { NonMemberHorseOwnerWithRelations, NonMemberHorseOwnerPartialWithRelations, NonMemberHorseOwnerOptionalDefaultsWithRelations } from './NonMemberHorseOwnerSchema'
+import { MemberWithRelationsSchema, MemberPartialWithRelationsSchema, MemberOptionalDefaultsWithRelationsSchema } from './MemberSchema'
+import type { MemberWithRelations, MemberPartialWithRelations, MemberOptionalDefaultsWithRelations } from './MemberSchema'
+import { RiderComboWithRelationsSchema, RiderComboPartialWithRelationsSchema, RiderComboOptionalDefaultsWithRelationsSchema } from './RiderComboSchema'
+import type { RiderComboWithRelations, RiderComboPartialWithRelations, RiderComboOptionalDefaultsWithRelations } from './RiderComboSchema'
 
 /////////////////////////////////////////
 // HORSE SCHEMA
