@@ -3,18 +3,10 @@ import { PhoneTypeSchema } from '../inputTypeSchemas/PhoneTypeSchema'
 import { TypeSchema } from '../inputTypeSchemas/TypeSchema'
 import { StatusSchema } from '../inputTypeSchemas/StatusSchema'
 import { StatusTypeSchema } from '../inputTypeSchemas/StatusTypeSchema'
-import type { RiderComboWithRelations } from './RiderComboSchema'
-import type { RiderComboPartialWithRelations } from './RiderComboSchema'
-import type { RiderComboOptionalDefaultsWithRelations } from './RiderComboSchema'
-import type { HorseWithRelations } from './HorseSchema'
-import type { HorsePartialWithRelations } from './HorseSchema'
-import type { HorseOptionalDefaultsWithRelations } from './HorseSchema'
-import { RiderComboWithRelationsSchema } from './RiderComboSchema'
-import { RiderComboPartialWithRelationsSchema } from './RiderComboSchema'
-import { RiderComboOptionalDefaultsWithRelationsSchema } from './RiderComboSchema'
-import { HorseWithRelationsSchema } from './HorseSchema'
-import { HorsePartialWithRelationsSchema } from './HorseSchema'
-import { HorseOptionalDefaultsWithRelationsSchema } from './HorseSchema'
+import { RiderComboWithRelationsSchema, RiderComboPartialWithRelationsSchema, RiderComboOptionalDefaultsWithRelationsSchema } from './RiderComboSchema'
+import type { RiderComboWithRelations, RiderComboPartialWithRelations, RiderComboOptionalDefaultsWithRelations } from './RiderComboSchema'
+import { HorseWithRelationsSchema, HorsePartialWithRelationsSchema, HorseOptionalDefaultsWithRelationsSchema } from './HorseSchema'
+import type { HorseWithRelations, HorsePartialWithRelations, HorseOptionalDefaultsWithRelations } from './HorseSchema'
 
 /////////////////////////////////////////
 // MEMBER SCHEMA
