@@ -1,29 +1,13 @@
 import { z } from 'zod';
 import { DivisionSchema } from '../inputTypeSchemas/DivisionSchema'
-import type { MemberWithRelations } from './MemberSchema'
-import type { MemberPartialWithRelations } from './MemberSchema'
-import type { MemberOptionalDefaultsWithRelations } from './MemberSchema'
-import type { HorseWithRelations } from './HorseSchema'
-import type { HorsePartialWithRelations } from './HorseSchema'
-import type { HorseOptionalDefaultsWithRelations } from './HorseSchema'
-import type { PointsWithRelations } from './PointsSchema'
-import type { PointsPartialWithRelations } from './PointsSchema'
-import type { PointsOptionalDefaultsWithRelations } from './PointsSchema'
-import type { ShowWithRelations } from './ShowSchema'
-import type { ShowPartialWithRelations } from './ShowSchema'
-import type { ShowOptionalDefaultsWithRelations } from './ShowSchema'
-import { MemberWithRelationsSchema } from './MemberSchema'
-import { MemberPartialWithRelationsSchema } from './MemberSchema'
-import { MemberOptionalDefaultsWithRelationsSchema } from './MemberSchema'
-import { HorseWithRelationsSchema } from './HorseSchema'
-import { HorsePartialWithRelationsSchema } from './HorseSchema'
-import { HorseOptionalDefaultsWithRelationsSchema } from './HorseSchema'
-import { PointsWithRelationsSchema } from './PointsSchema'
-import { PointsPartialWithRelationsSchema } from './PointsSchema'
-import { PointsOptionalDefaultsWithRelationsSchema } from './PointsSchema'
-import { ShowWithRelationsSchema } from './ShowSchema'
-import { ShowPartialWithRelationsSchema } from './ShowSchema'
-import { ShowOptionalDefaultsWithRelationsSchema } from './ShowSchema'
+import { MemberWithRelationsSchema, MemberPartialWithRelationsSchema, MemberOptionalDefaultsWithRelationsSchema } from './MemberSchema'
+import type { MemberWithRelations, MemberPartialWithRelations, MemberOptionalDefaultsWithRelations } from './MemberSchema'
+import { HorseWithRelationsSchema, HorsePartialWithRelationsSchema, HorseOptionalDefaultsWithRelationsSchema } from './HorseSchema'
+import type { HorseWithRelations, HorsePartialWithRelations, HorseOptionalDefaultsWithRelations } from './HorseSchema'
+import { PointsWithRelationsSchema, PointsPartialWithRelationsSchema, PointsOptionalDefaultsWithRelationsSchema } from './PointsSchema'
+import type { PointsWithRelations, PointsPartialWithRelations, PointsOptionalDefaultsWithRelations } from './PointsSchema'
+import { ShowWithRelationsSchema, ShowPartialWithRelationsSchema, ShowOptionalDefaultsWithRelationsSchema } from './ShowSchema'
+import type { ShowWithRelations, ShowPartialWithRelations, ShowOptionalDefaultsWithRelations } from './ShowSchema'
 
 /////////////////////////////////////////
 // RIDER COMBO SCHEMA
